Compute everything in float64
This simplifies type conversions

+++ pkg/lm/projection.go
@@ -6,19 +6,19 @@ func DimetricProjection() Matrix3 {
 	alpha := math.Pi / 6
 	beta := math.Pi / 4
 
-	cosAlpha := float32(math.Cos(alpha))
-	sinAlpha := float32(math.Sin(alpha))
+	cosAlpha := math.Cos(alpha)
+	sinAlpha := math.Sin(alpha)
 
-	cosBeta := float32(math.Cos(beta))
-	sinBeta := float32(math.Sin(beta))
+	cosBeta := math.Cos(beta)
+	sinBeta := math.Sin(beta)
 
-	rotateX := NewMatrix3([9]float32{
+	rotateX := NewMatrix3([9]float64{
 		1, 0, 0,
 		0, cosAlpha, sinAlpha,
 		0, -sinAlpha, cosAlpha,
 	})
 
-	rotateY := NewMatrix3([9]float32{
+	rotateY := NewMatrix3([9]float64{
 		cosBeta, 0, -sinBeta,
 		0, 1, 0,
 		sinBeta, 0, cosBeta,
