+++ pkg/lm/vector4.go
@@ -1,10 +1,10 @@
 package lm
 
 type Vector4 struct {
-	X, Y, Z, W float32
+	X, Y, Z, W float64
 }
 
-func Vec4(x, y, z, w float32) Vector4 {
+func Vec4(x, y, z, w float64) Vector4 {
 	return Vector4{
 		X: x,
 		Y: y,
@@ -13,7 +13,7 @@ func Vec4(x, y, z, w float32) Vector4 {
 	}
 }
 
-func (lhs Vector4) MulScalar(rhs float32) Vector4 {
+func (lhs Vector4) MulScalar(rhs float64) Vector4 {
 	x := lhs.X * rhs
 	y := lhs.Y * rhs
 	z := lhs.Z * rhs
@@ -21,7 +21,7 @@ func (lhs Vector4) MulScalar(rhs float32) Vector4 {
 	return Vec4(x, y, z, w)
 }
 
-func (lhs Vector4) ClampScalar(min, max float32) Vector4 {
+func (lhs Vector4) ClampScalar(min, max float64) Vector4 {
 	x := Clamp(lhs.X, min, max)
 	y := Clamp(lhs.Y, min, max)
 	z := Clamp(lhs.Z, min, max)
